clean up bid/registration flow queryrenderers
