+++ src/lib/Containers/RegistrationFlow.tsx
@@ -1,10 +1,13 @@
 import React from "react"
 import { ViewProperties } from "react-native"
 import NavigatorIOS from "react-native-navigator-ios"
-import { createFragmentContainer, graphql } from "react-relay"
+import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 
 import { RegistrationFlow_me } from "__generated__/RegistrationFlow_me.graphql"
 import { RegistrationFlow_sale } from "__generated__/RegistrationFlow_sale.graphql"
+import { RegistrationFlowQuery } from "__generated__/RegistrationFlowQuery.graphql"
+import { defaultEnvironment } from "lib/relay/createEnvironment"
+import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import { TimeOffsetProvider } from "../Components/Bidding/Context/TimeOffsetProvider"
 import { RegistrationScreen } from "../Components/Bidding/Screens/Registration"
 
@@ -31,7 +34,7 @@ class RegistrationFlow extends React.Component<RegistrationFlowProps> {
   }
 }
 
-export default createFragmentContainer(RegistrationFlow, {
+export const RegistrationFlowFragmentContainer = createFragmentContainer(RegistrationFlow, {
   sale: graphql`
     fragment RegistrationFlow_sale on Sale {
       ...Registration_sale
@@ -43,3 +46,27 @@ export default createFragmentContainer(RegistrationFlow, {
     }
   `,
 })
+
+export const RegistrationFlowRenderer: React.SFC<{ saleID: string }> = ({ saleID }) => {
+  return (
+    <QueryRenderer<RegistrationFlowQuery>
+      environment={defaultEnvironment}
+      query={graphql`
+        query RegistrationFlowQuery($saleID: String!) {
+          sale(id: $saleID) {
+            name
+            ...RegistrationFlow_sale
+          }
+          me {
+            ...RegistrationFlow_me
+          }
+        }
+      `}
+      cacheConfig={{ force: true }} // We want to always fetch latest sale registration status, CC info, etc.
+      variables={{
+        saleID,
+      }}
+      render={renderWithLoadProgress(RegistrationFlowFragmentContainer)}
+    />
+  )
+}
