+++ src/lib/Containers/__tests__/BidFlow-tests.tsx
@@ -2,9 +2,8 @@ import React from "react"
 import "react-native"
 import * as renderer from "react-test-renderer"
 
-import BidFlow from "../BidFlow"
-
 import { Theme } from "@artsy/palette"
+import { BidFlowFragmentContainer } from "../BidFlow"
 
 jest.mock("tipsi-stripe", () => ({ setOptions: jest.fn() }))
 
@@ -45,7 +44,7 @@ it("renders properly", () => {
   const bg = renderer
     .create(
       <Theme>
-        <BidFlow me={Me as any} sale_artwork={SaleArtwork as any} />
+        <BidFlowFragmentContainer me={Me as any} sale_artwork={SaleArtwork as any} />
       </Theme>
     )
     .toJSON()
